requirements.txt
azure-common==1.1.27
azure-core==1.18.0
azure-data-tables==12.1.0
azure-identity==1.6.1
azure-mgmt-automation==1.0.0
azure-mgmt-core==1.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
Flask-WTF==0.15.1
idna==3.2
importlib-metadata==4.8.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msal==1.14.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.1
portalocker==1.7.1
pycparser==2.20
PyJWT==2.1.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
